+++ browser/lib/src/store.ts
@@ -326,7 +326,7 @@ export class Store {
     if (opts.setLoading) {
       const newR = new Resource<C>(subject);
       newR.loading = true;
-      this.addResources(newR);
+      this.addResources(newR, { skipCommitCompare: true });
     }
 
     // Use WebSocket if available, else use HTTP(S)
@@ -357,7 +357,7 @@ export class Store {
         },
       );
 
-      this.addResources(createdResources);
+      this.addResources(createdResources, { skipCommitCompare: true });
     }
 
     return this.resources.get(subject)!;