+++ browser/CHANGELOG.md
@@ -19,10 +19,12 @@ This changelog covers all five packages, as they are (for now) updated as a whol
 - Redesigned the ontology page.
 - Moved the resource context menu to the top of the page.
 - [#861](https://github.com/atomicdata-dev/atomic-server/issues/861) Fix long usernames overflowing on the share page.
+- [#906](https://github.com/atomicdata-dev/atomic-server/issues/906) Reset changes after clicking the cancel button in a form or navigating away.
 
 ### @tomic/lib
 
 - Added `LocalChange` event to `Resource`.
+- Added `resource.refresh()` method.
 
 ### @tomic/react